Add setup and teardown to test runner and make use of that in tests
Hopefully this will make creating test case 5 a bit easier.
It should also help to extract the system logic from the test cases. As this is a simulator built to test how the card will behave when running these tests, the boundary between system behaviour and test behaviour has gotten a bit muddied. This is probably all overkill but its been an interesting exercise.