Add positive and negative playwright tests for test cases 1-4
This is a first pass so locators and actions are duplicated right now.

Also test cases must be run in order which I don't like, but the SUT is stateful and its 'tests' run one after the other, so the playwright tests must match that. The test groups can be run in any order though.