Working ISR
Note: After a couple of days trying to track down hard faults, turns out it was down to the type of interrupt used.

From what I understand "WCH-Interrupt-fast" is expecting the ISR to be stack safe - that is to say, don't call anything that touches the stack (library calls, functions, variables etc). Added the config entry to use stack safe ISR.